requirements.txt
numpy
scipy
soundfile
librosa
torchaudio
fastapi
uvicorn[standard]
aiofiles
python-dotenv
websockets
sounddevice
transformers
